refactor: 패키지 구조 변경 및 인터페이스 사용
