--- src/main/java/com/avatar/search/batch/domain/ProductRepository.java
@@ -1,6 +1,0 @@
-package com.avatar.search.batch.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
