--- src/main/java/com/avatar/search/batch/dto/BatchItemDTO.java
@@ -1,7 +1,0 @@
-package com.avatar.search.batch.dto;
-
-import java.util.List;
-
-public record BatchItemDTO(HotelDTO hotel,
-                           List<NaverApiItemResponse> responses) {
-}
