--- src/main/java/com/avatar/search/batch/dto/HotelDTO.java
@@ -1,4 +1,0 @@
-package com.avatar.search.batch.dto;
-
-public record HotelDTO (Long id, String name) {
-}
