--- src/main/java/com/avatar/search/batch/dto/NaverApiResponse.java
@@ -1,6 +1,0 @@
-package com.avatar.search.batch.dto;
-
-import java.util.List;
-
-public record NaverApiResponse(String lastBuildDate, long total, int start, int display, List<NaverApiItemResponse> items) {
-}
