+++ src/main/java/com/avatar/search/client/DataClient.java
@@ -1,0 +1,8 @@
+package com.avatar.search.client;
+
+import com.avatar.search.dto.ClientDataRequest;
+import com.avatar.search.dto.ClientDataResponse;
+
+public interface DataClient {
+    ClientDataResponse getData(ClientDataRequest request);
+}
