+++ src/main/java/com/avatar/search/domain/hotel/Area.java
@@ -1,0 +1,36 @@
+package com.avatar.search.domain.hotel;
+
+import com.avatar.search.common.domain.BaseEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+
+import java.util.Objects;
+
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Area extends BaseEntity {
+
+    @Id
+    private String code;
+    private String name;
+    private Integer orderSeq;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Area area = (Area) o;
+        return Objects.equals(code, area.code);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(code);
+    }
+}
