+++ src/main/java/com/avatar/search/domain/hotel/Hotel.java
@@ -1,0 +1,52 @@
+package com.avatar.search.domain.hotel;
+
+import com.avatar.search.common.domain.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+import java.util.Objects;
+
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Hotel extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+    private String address;
+    private String phoneNumber;
+    private String homepage;
+    private Integer rooms;
+
+    @Enumerated(value = EnumType.STRING)
+    private Classify classification;
+
+    @Enumerated(value = EnumType.STRING)
+    private Grade grade;
+
+    private LocalDate gradingDate;
+
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.PERSIST)
+    @JoinColumn(name = "area_code", foreignKey = @ForeignKey(name = "hotel_area_code_fk"))
+    private Area area;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Hotel hotel = (Hotel) o;
+        return id.equals(hotel.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+}
