+++ src/main/java/com/avatar/search/domain/hotel/HotelRepository.java
@@ -1,0 +1,6 @@
+package com.avatar.search.domain.hotel;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HotelRepository extends JpaRepository<Hotel, Long> {
+}
