+++ src/main/java/com/avatar/search/domain/item/ItemRepository.java
@@ -1,0 +1,6 @@
+package com.avatar.search.domain.item;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
+}
