+++ src/main/java/com/avatar/search/dto/BatchItemDTO.java
@@ -1,0 +1,7 @@
+package com.avatar.search.dto;
+
+import java.util.List;
+
+public record BatchItemDTO(HotelDTO hotel,
+                           List<ClientDataItemResponse> responses) {
+}
