+++ src/main/java/com/avatar/search/dto/ClientDataResponse.java
@@ -1,0 +1,6 @@
+package com.avatar.search.dto;
+
+import java.util.List;
+
+public record ClientDataResponse (String lastDate, long total, int page, int display, List<ClientDataItemResponse> items){
+}
