+++ src/main/java/com/avatar/search/dto/HotelDTO.java
@@ -1,0 +1,4 @@
+package com.avatar.search.dto;
+
+public record HotelDTO (Long id, String name) {
+}
