--- src/main/java/com/avatar/search/hotel/domain/HotelRepository.java
@@ -1,6 +1,0 @@
-package com.avatar.search.hotel.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HotelRepository extends JpaRepository<Hotel, Long> {
-}
