+++ src/main/java/com/avatar/search/infrastructure/dto/NaverApiSort.java
@@ -1,0 +1,17 @@
+package com.avatar.search.infrastructure.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum NaverApiSort {
+    sim("sim", "정확도순으로 내림차순"),
+    date("date", "정확도순으로 내림차순"),
+    asc("asc", "정확도순으로 내림차순"),
+    dsc("dsc", "정확도순으로 내림차순");
+
+    private final String code;
+    private final String desc;
+
+}
